Factor out an openedResource type.
